tech(TTB-060): Moved delay functions to general util file
