+++ client/src/components/Game/GameRobot.jsx
@@ -5,8 +5,8 @@ import Player from './Player';
 import { Link } from 'react-router-dom';
 import { MdHome, MdReplay, MdOutlineSettings } from 'react-icons/md';
 import './game.css';
-import { calculateWinner, compareArrays, removeDuplicatesFromArray, delay, checkIfWinningMove, calculateSurroundingCoordinates, checkIfSetUpWinningMove } from '../../utils/helper/gameUtils';
-
+import { calculateWinner, compareArrays, removeDuplicatesFromArray, checkIfWinningMove, calculateSurroundingCoordinates, checkIfSetUpWinningMove } from '../../utils/helper/gameUtils';
+import { delay } from '../../utils/delay';
 
 function GameRobot(props) {
   const playerOne = new Player('❌');
