+++ client/src/utils/delay.js
@@ -1,0 +1,5 @@
+export const delay = (duration) => {
+    return new Promise((resolve) => {
+      setTimeout(() => resolve(), duration);
+    });
+}
